CARLI-528 Connect the notification templates from the notification screen to the modals.
Making a change to a template in the manage modal didn't show up in the draft notification modal. Rather than adding lots of callbacks we just load the templates in the main screen controller and pass it in to both modals so it is bound between them and automatically updates. Be careful to discard changes when closing the manage modal so they don't show up on the edit draft modal.

The draft notification modal now might have the templates passed in (notification screen) or it might not (everywhere else in the app). So it checks whether it got some via the isolate scope and loads the templates if it needs them.