CARLI-598 A bunch of cleanup in the Gruntfile to get parallel builds of carli and vendor working.
The grunt:serve task serves the build/ directory, and carliApp/ and vendorApp/ each have everything in them that they need to run. Except I still need to make separate Sass builds for each, and also copy the images.

We need to decide how we want to serve the apps in development now, and fix up the connect task to reflect that. Also the serve and build tasks need to be adjusted to build all or some of the apps.