CARLI-502 Add a red border and warning message when date picker value is invalid.
The implementation of the date picker we are using causes it to ignore invalid dates (they don't get set on the hg-model at all), so we don't get a failing validation message because the invalid value is never sent. This solution does not prevent the form from saving, but it does warn the user that their invalid date will not be saved.