CARLI-640 set up empty offerings for new OTP products for all libraries.
Generate empty offerings for all libraries when setting up the new product form for a one-time-purchase so that prices can be entered. Only save offerings that have prices entered. When loading a product, make sure to generate empty offerings for libraries for which the product does not already have an offering. If saving a new product, create all the offerings once the product is created. If saving an existing product, update the offerings that already existed but create the new offerings that were initialized and had prices entered. This also handles creating empty offerings (with no prices) for all libraries for new products that are not one-time-purchase products.