CARLI-754 Add toggle to product headers.
Had to do some CSS hackery to make the affixed headers stack under the non-affixed headers. The product headers that are still collapsed don't get position fixed, but if there was an expanded one at the top and you scroll down that header will stay with you.

Known issue: the purple bar goes all the way to the right-hand side of the window. The header is expanding to fill the container and then when it gets position: fixed it fills the whole window. It happens to have a correct left positioning though.