CARLI-755 Use angular events to update drafts and sent lists when they change.
Removed old unused code that was previously used to connect these lists which got changed / broken recently.
Since the lists of notifications are stored and managed by the list components themselves instead of the parent (for filtering purposes) there needed to be a way to signal that the lists updated to each other. Including the parent screen when a new notification is drafted.