CARLI-994 Use ID's to prevent latent offering row appending.
I think that's what was happening anyway - the data would load and the jQuery would fire to build the grid and append the row. And since both screens were using '.pricing-grid' as the selector it would still add them even though you had moved to a different screen. Using different ID's help this.